Move NicetyRow into its own file
